empty list check

+++ src/rag_system/retriever.py
@@ -153,6 +153,9 @@ class VectorStore:
         #     print(f"Text: {documents[result_item.index]}\n")
 
     def remove_by_ids(self, ids: list[str]):
+        if not len(ids):
+            return []
+
         docs = self.collection.delete(ids=ids)
         return docs
